+++ lib/src/scidart/signal/windows/flattop.dart
@@ -36,6 +36,7 @@ import 'general_cosine.dart';
 /// var w = flattop(M);
 /// ```
 Array flattop(int M, {bool sym = true}) {
-  var a = Array([0.21557895, 0.41663158, 0.277263158, 0.083578947, 0.006947368]);
+  var a =
+      Array([0.21557895, 0.41663158, 0.277263158, 0.083578947, 0.006947368]);
   return generalCosine(M, a, sym: sym);
 }