+++ lib/src/scidart/signal/windows/nuttall.dart
@@ -30,5 +30,6 @@ import 'general_cosine.dart';
 /// var w = nuttall(M);
 /// ```
 Array nuttall(int M, {bool sym = true}) {
-  return generalCosine(M, Array([0.3635819, 0.4891775, 0.1365995, 0.0106411]), sym: sym);
+  return generalCosine(M, Array([0.3635819, 0.4891775, 0.1365995, 0.0106411]),
+      sym: sym);
 }