+++ test/numdart/spaces/zeros_test.dart
@@ -15,4 +15,3 @@ void main() {
     expect(n.length, nExpect.length);
   });
 }
-