+++ test/scidart/special/array_bessel_i0_test.dart
@@ -9,7 +9,7 @@ void main() {
     print(i0);
 
     var i0Expec =
-    Array([1.00000000e+00, 1.07375171e+42, 1.00000000e+00, 1.07375171e+42]);
+        Array([1.00000000e+00, 1.07375171e+42, 1.00000000e+00, 1.07375171e+42]);
 
     arrayTruncateEachElement(i0, 5);
     arrayTruncateEachElement(i0Expec, 5);